pyproject.toml
[project]
name = "starborn-eduflow"
version = "0.1.0"
description = "EduFlow AI Education Platform — Backend API"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.13.0",
    "anthropic>=0.93.0",
    "asyncpg>=0.30.0",
    "bcrypt>=4.0.0",
    "bleach>=6.0.0",
    "celery[redis]>=5.3.0",
    "email-validator>=2.0.0",
    "fastapi>=0.115.0",
    "greenlet>=3.0.0",
    "openai>=2.31.0",
    "pydantic>=2.12.5",
    "pyjwt>=2.8.0",
    "python-dotenv>=1.0.0",
    "slowapi>=0.1.9",
    "sqlalchemy[asyncio]>=2.0.0",
    "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
# Linux-only optimizatsiyalar (Docker / Railway / production)
# Docker da: uv sync --extra linux
linux = [
    "uvloop>=0.19.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]
